pyproject.toml
[project]
name = "zerobypass"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest-playwright>=0.7.2",
    "rich>=14.3.2",
]
